Module Cwe_checker_core.Cwe_332

How the check works

We check whether the program calls rand but not srand.

False Positives

None known

False Negatives

It is not checked whether srand gets called before rand

val name : string

val version : string

val check_cwe : Bap.Std.program Bap.Std.term -> 'a -> 'b -> string list list -> 'c -> unit