Module Cwe_checker_core.Cwe_367

How the check works

For pairs of (check-call, use-call), configurable in config.json, we check whether a function may call the check-call before the use-call.

False Positives

The check-call and the use-call may access different, unrelated resources (e. g. different files).

False Negatives

If the check-call and the use-call happen in different functions it will not be found by the check.

val name : string

val version : string

val check_cwe : Bap.Std.program Bap.Std.term -> Bap.Std.project -> Bap.Std.word Bap.Std.Tid.Map.t -> string list list -> string list -> unit