Module Cwe_checker_core.Cwe_426

How the check works

We check whether a function that calls a privilege-changing function (configurable in config.json) also calls system().

False Positives

If the call to system() happens before the privilege-changing function, the call may not be used for privilege escalation

False Negatives

If the calls to the privilege-changing function and system() happen in different functions, the calls will not be flagged as a CWE-hit.

This check only finds potential privilege escalation bugs, but other types of bugs can also be triggered by untrusted search paths.

val name : string

val version : string

val check_cwe : Bap.Std.program Bap.Std.term -> Bap.Std.project -> Bap.Std.word Bap.Std.Tid.Map.t -> string list list -> string list -> unit