Module Cwe_checker_core.Cwe_476

How the check works

Using dataflow analysis we search for an execution path where a memory access using the return value of a symbol happens before the return value is checked through a conditional jump instruction.

Note that the check relies on Bap-generated stubs to identify return registers of the checked functions. Therefore it only works for functions for which Bap generates these stubs.

Parameters configurable in config.json

strict_call_policy={true, false}: Determines behaviour on call and return instructions. If false, we assume that the callee, resp. the caller on a return instruction, checks all unchecked values still contained in parameter registers. If true, every unchecked value on a call or return instruction gets reported.

strict_mem_policy={true, false}: Determines behaviour on writing an unchecked return value to a memory region other than the stack. If true, these instances get reported. Depending on the coding style, this can lead to a lot false positives if return values are only checked after writing them to their target destination. If false, these instances do not get reported, which in turn can lead to false negatives.

max_steps=<num>: Max number of steps for the dataflow fixpoint algorithm.

Symbols configurable in config.json

The symbols are the functions whose return values are assumed to be potential NULL pointers.

False Positives

If strict_mem_policy is set to true, writing a return value to memory other than the stack gets reported even if a NULL pointer check happens right afterwards.

The check has no knowledge about the actual number of parameters that an extern function call takes. This can lead to false positives if strict_call_policy is set to true.

False Negatives

We do not check whether an access to a potential NULL pointer happens regardless of a prior check.

We do not check whether the conditional jump instruction checks specifically for the return value being NULL or something else

For functions with more than one return value we do not distinguish between the return values.

If strict_mem_policy is set to false, unchecked return values that are saved somewhere other than the stack may be missed.

The check has no knowledge about the actual number of parameters that an extern function call takes. This can lead to false negatives, especially if function parameters are passed on the stack.

val name : string

val version : string

val check_cwe : Bap.Std.program Bap.Std.term -> Bap.Std.project -> Bap.Std.word Bap.Std.Tid.Map.t -> string list list -> string list -> unit