Module Cwe_checker_core.Cwe_560

How the check works

This check looks for umask calls and checks if they have a reasonable value, i.e. smaller than a certain value, currently set to 1000 and greater than a reasonable value for umask, currently set to 100.

False Positives

The current version considers all immediate values of an umask callsite's basic block. It does not check whether the value is an input to the call or not.

False Negatives

If the input to umask is not defined in the basic block before the call, the check will not see it.

Calls where the input is not an immediate value but a variable are not examined.

val name : string

val version : string

val check_cwe : Bap.Std.program Bap.Std.term -> Bap.Std.project -> Bap.Std.word Bap.Std.Tid.Map.t -> string list list -> string list -> unit

module Private : sig ... end