Module Cwe_checker_core.Mem_region

type 'a t

val bin_t : 'a Core_kernel.Bin_prot.Type_class.t -> 'a t Core_kernel.Bin_prot.Type_class.t

val bin_read_t : 'a Core_kernel.Bin_prot.Read.reader -> 'a t Core_kernel.Bin_prot.Read.reader

val __bin_read_t__ : 'a Core_kernel.Bin_prot.Read.reader -> (int -> 'a t) Core_kernel.Bin_prot.Read.reader

val bin_reader_t : 'a Core_kernel.Bin_prot.Type_class.reader -> 'a t Core_kernel.Bin_prot.Type_class.reader

val bin_size_t : 'a Core_kernel.Bin_prot.Size.sizer -> 'a t Core_kernel.Bin_prot.Size.sizer

val bin_write_t : 'a Core_kernel.Bin_prot.Write.writer -> 'a t Core_kernel.Bin_prot.Write.writer

val bin_writer_t : 'a Core_kernel.Bin_prot.Type_class.writer -> 'a t Core_kernel.Bin_prot.Type_class.writer

val bin_shape_t : Core_kernel.Bin_prot.Shape.t -> Core_kernel.Bin_prot.Shape.t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val t_of_sexp : (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a t

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

val empty : unit -> 'a t

Get an empty memory region-

val add : 'a t -> 'a -> pos:Bap.Std.Bitvector.t -> size:Bap.Std.Bitvector.t -> 'a t

Add an element to the memory region. If the element intersects existing elements, the non-overwritten part gets marked as Error

val remove : 'a t -> pos:Bap.Std.Bitvector.t -> size:Bap.Std.Bitvector.t -> 'a t

Mark the memory region between pos (included) and pos+size (excluded) as empty. If elements get partially removed, mark the non-removed parts as Error

val get : 'a t -> Bap.Std.Bitvector.t -> ('a * Bap.Std.Bitvector.t, unit) Core_kernel.Result.t Core_kernel.Option.t

Returns the element and its size at position pos or None, when there is no element at that position. If pos intersects an element but does not match its starting position, it returns Some(Error(())).

val merge : 'a t -> 'a t -> data_merge:('a -> 'a -> ('a, 'b) Core_kernel.result option) -> 'a t

Merge two memory regions. Elements with the same position and size get merged using data_merge, other intersecting elements get marked as Error. Note that data_merge may return None (to remove the elements from the memory region) or Some(Error(_)) to mark the merged element as error.

val equal : 'a t -> 'a t -> data_equal:('a -> 'a -> bool) -> bool

Check whether two memory regions are equal.

val mark_error : 'a t -> pos:Bap.Std.Bitvector.t -> size:Bap.Std.Bitvector.t -> 'a t

Mark an area in the mem_region as containing errors.