Module Cwe_checker_core__Cconv

val get_json : unit -> Yojson.Basic.t

Returns a json object for registers.json. If not yet in memory, the json file is read.

val is_callee_saved : Bap.Std.Var.t -> Bap.Std.Project.t -> Core_kernel.Bool.t

Returns whether a variable is callee saved according to the calling convention of the target architecture. Should only used for calls to functions outside of the program, not for calls between functions inside the program.

val is_parameter_register : Bap.Std.Var.t -> Bap.Std.Project.t -> Core_kernel.Bool.t

Returns whether a variable may be used to pass parameters to a function. This depends on the calling convention of the target architecture and should only be used for extern function calls.

val is_return_register : Bap.Std.Var.t -> Bap.Std.Project.t -> Core_kernel.Bool.t

Returns whether a variable may be used for return values of function calls. This depends on the calling convention of the target architecture and should only be used for extern function calls.

val get_supported_architectures : unit -> string list

Returns a string list of supported architectures from the registers.json.

val infer_bin_format_from_symbols : Bap.Std.Project.t -> string

Infers the binary format using the file's symbol table.

val extract_bin_format : Bap.Std.Project.t -> string

Returns the binary format by either an objdump call or via the file's symbol table.

val get_register_list : Bap.Std.Project.t -> string -> string list

Returns a list of registers based on the file's binary format, architecture, calling_convention and context (e.g. callee saved, parameter etc.)