Module Cwe_checker_core__Main

This module defines the main driving function for the cwe_checker plugin in BAP.

type cwe_module = {

cwe_func : Bap.Std.program Bap.Std.term -> Bap.Std.project -> Bap.Std.word Bap.Std.Tid.Map.t -> string list list -> string list -> unit;

name : string;

version : string;

requires_pairs : bool;

has_parameters : bool;

}

val version : Core_kernel.String.t

val print_version : unit -> unit

prints the version number

val known_modules : cwe_module Core_kernel.List.t

val cmdline_flags : (Core_kernel.String.t * Core_kernel.String.t) Core_kernel.List.t

val cmdline_params : (Core_kernel.String.t * Core_kernel.String.t) Core_kernel.List.t

val print_module_versions : unit -> unit

val check_valid_module_list : string list -> unit

val print_help_message : unit -> unit

prints the help message

val execute_cwe_module : cwe_module -> Yojson.Basic.t -> Bap.Std.project -> Bap.Std.program Bap.Std.term -> Bap.Std.word Bap.Std.Tid.Map.t -> unit

Executes one CWE module

val partial_run : Yojson.Basic.t -> Bap.Std.project -> Bap.Std.program Bap.Std.term -> Bap.Std.word Bap.Std.Tid.Map.t -> string list -> unit

Only runs checks on CWE module specified by user.

val full_run : Yojson.Basic.t -> Bap.Std.project -> Bap.Std.program Bap.Std.term -> Bap.Std.word Bap.Std.Tid.Map.t -> unit

Runs checks on all supported CWE modules.

val build_output_path : string -> string

val main : Core_kernel.Bool.t Core_kernel.String.Map.t -> Core_kernel.String.t Core_kernel.String.Map.t -> Bap.Std.Project.t -> unit

The main function drives the execution of the cwe_checker plugin in BAP. The command line arguments are passed as maps from their name to to their values (Bool.t for flags, String.t for other arguments) to this function.