Module Cwe_checker_core__Mem_region

contains an abstract memory region data type where you can assign arbitrary data to locations inside the memory regions. A memory region has no fixed size, so it can be used for memory regions of variable size like arrays or stacks.

TODO: Right now this data structure is unsuited for elements that get only partially loaded.

type 'a t

include Core_kernel.Bin_prot.Binable.S1 with type 'a Cwe_checker_core__Mem_region.t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

include Ppx_sexp_conv_lib.Sexpable.S1 with type 'a Cwe_checker_core__Mem_region.t := 'a t

val t_of_sexp : (Sexplib0__.Sexp.t -> 'a) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0__.Sexp.t) -> 'a t -> Sexplib0__.Sexp.t

val empty : unit -> 'a t

Get an empty memory region-

val add : 'a t -> 'a -> pos:Bap.Std.Bitvector.t -> size:Bap.Std.Bitvector.t -> 'a t

Add an element to the memory region. If the element intersects existing elements, the non-overwritten part gets marked as Error

val remove : 'a t -> pos:Bap.Std.Bitvector.t -> size:Bap.Std.Bitvector.t -> 'a t

Mark the memory region between pos (included) and pos+size (excluded) as empty. If elements get partially removed, mark the non-removed parts as Error

val get : 'a t -> Bap.Std.Bitvector.t -> ('a * Bap.Std.Bitvector.t, unit) Core_kernel.Result.t Core_kernel.Option.t

Returns the element and its size at position pos or None, when there is no element at that position. If pos intersects an element but does not match its starting position, it returns Some(Error(())).

val merge : 'a t -> 'a t -> data_merge:('a -> 'a -> ('a, 'b) Core_kernel.result option) -> 'a t

Merge two memory regions. Elements with the same position and size get merged using data_merge, other intersecting elements get marked as Error. Note that data_merge may return None (to remove the elements from the memory region) or Some(Error(_)) to mark the merged element as error.

val equal : 'a t -> 'a t -> data_equal:('a -> 'a -> bool) -> bool

Check whether two memory regions are equal.

val mark_error : 'a t -> pos:Bap.Std.Bitvector.t -> size:Bap.Std.Bitvector.t -> 'a t

Mark an area in the mem_region as containing errors.

val map_data : 'a t -> f:('a -> 'b) -> 'b t

Map the contained data to new values.

val list_data : 'a t -> 'a Core_kernel.List.t

List the contained data (ignoring error values).

val list_data_pos : 'a t -> (Bap.Std.Bitvector.t * 'a) Core_kernel.List.t

List the contained data (ignoring error values) together with their positions.