Module Cwe_checker_core__Serde_json

This module allows the creation of SerdeJson objects that can be deserialized to the corresponding data type in Rust.

Note that this is not optimized for speed, extensive usage could lead to measureable slowdown.

type t

val build_null : unit -> t

Build a Json Null object

val build_bool : Core_kernel.Bool.t -> t

Build a Json boolean object

val build_number : int -> t

Build a Json number object

val build_string : Core_kernel.String.t -> t

Build a Json string object

val build_array : t Core_kernel.List.t -> t

Build a Json array object from a list of Json objects

val build_object : (Core_kernel.String.t * t) Core_kernel.List.t -> t

Build a Json object from a list of key-value-pairs

val to_string : t -> Core_kernel.String.t

Get the Json string corresponding to a Json object

val of_var_type : Bap.Std.Bil.Types.typ -> t

val of_var : Bap.Std.Var.t -> t

val of_cast_type : Bap.Std.Bil.Types.cast -> t

val of_binop_type : Bap.Std.Bil.Types.binop -> t

val of_unop_type : Bap.Std.Bil.Types.unop -> t

val of_endianness : Bap.Std.Bitvector.endian -> t

val of_bitvector : Bap.Std.Bitvector.t -> t

val of_exp : Bap.Std.Exp.t -> t

val of_tid : Bap.Std.Tid.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_def : Bap.Std.Def.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_jmp_label : Bap.Std.Label.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_call : Bap.Std.Call.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_jmp_kind : Bap.Std.jmp_kind -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_jmp : Bap.Std.Jmp.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_blk : Bap.Std.Blk.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_sub : Bap.Std.Sub.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_extern_symbol : Cwe_checker_core.Symbol_utils.extern_symbol -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_program : Bap.Std.Program.t -> Cwe_checker_core.Symbol_utils.extern_symbol Core_kernel.List.t -> Bap.Std.Tid.t Core_kernel.List.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t

val of_project : Bap.Std.Project.t -> Cwe_checker_core.Symbol_utils.extern_symbol Core_kernel.List.t -> Bap.Std.Tid.t Core_kernel.List.t -> Bap.Std.word Bap.Std.Tid.Map.t -> t