Module Cwe_checker_core__Type_inference

module PointerTargetInfo : sig ... end

The PointerTargetInfo contains knowledge about the offset and the alignment of a pointer into a memory area. Here the alignment is always considered relative offset zero of the target memory area.

module Register : sig ... end

The Register.t type. A register holds either arbitrary data or a pointer to some memory region. We do track possible targets of the pointer as follows:

module TypeInfo : sig ... end

The TypeInfo module. A TypeInfo.t structure holds a list of registers with known type information (see Register.t type) and known type information for values on the stack.

val type_info_tag : TypeInfo.t Bap.Std.Value.tag

A tag for TypeInfo.t, so that we can annotate basic blocks with known type information using bap tags.

val compute_pointer_register : Bap.Std.Project.t -> Bap.Std.Project.t

Computes TypeInfo for the given project. Adds tags to each block containing the TypeInfo at the start of the block.

val print_type_info_tags : project:Bap.Std.Project.t -> tid_map:Bap.Std.word Bap.Std.Tid.Map.t -> unit

Print type info tags. TODO: If this should be used for more than debug purposes, then the output format should be refactored accordingly.

val update_type_info : Bap.Std.Blk.elt -> TypeInfo.t -> sub_tid:Bap.Std.Tid.t -> project:Bap.Std.Project.t -> TypeInfo.t

Updates the type info for a single element (Phi/Def/Jmp) of a block. Input is the type info before execution of the element, output is the type info after execution of the element. sub_tid is the Tid of the current function which is internally used to mark which pointers point to the current stack frame.

module Private : sig ... end