Module Cwe_checker_core__Type_inference.PointerTargetInfo

The PointerTargetInfo contains knowledge about the offset and the alignment of a pointer into a memory area. Here the alignment is always considered relative offset zero of the target memory area.

type t = {

offset : (Bap.Std.Bitvector.t, unit) Core_kernel.Result.t Core_kernel.Option.t;

alignment : (int, unit) Core_kernel.Result.t Core_kernel.Option.t;

}

include Core_kernel.Bin_prot.Binable.S with type PointerTargetInfo.t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> int

include Ppx_sexp_conv_lib.Sexpable.S with type PointerTargetInfo.t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t